investigation(btree): 完成数据丢失根本原因调查并添加详细报告
根本发现：1076 个 key 阈值处数据丢失的根本原因是
页面内 keys 完全无序，导致 SearchKey (二分查找) 返回错误位置，
形成恶性循环。

关键证据：
- Page 92 keys: key-0, key-1, key-10, key-100, key-1000... (完全无序)
- Page 16 范围: First=key-245, Last=key-28 (颠倒)
- 1076 个 key 只有 972 个在子页面中 (丢失 104 个)

修改内容：
- offheap_adapter.go: 添加 UpdateIndexEntry extraChild 处理的详细注释
- materialize.go: 修复 EncodeChildWithVersion 类型转换 (uint32->uint64)
- 新增调查报告: docs/09_code-review/btree-data-loss-root-cause-report.md
- 新增测试: binary_search_integrity_test.go, present_keys_test.go,
            page_content_test.go (用于定位精确失败阈值 1076)

建议后续方案：
- 方案 A (短期): InsertLeafEntry 添加有序性验证
- 方案 B (中期): 检测无序时重建页面
- 方案 C (长期): 重新设计插入逻辑

相关 Issue: #18

+++ internal/infrastructure/storage/btree/offheap/materialize.go
@@ -84,7 +84,7 @@ func (m *OffHeapMaterializer) MaterializeIndexPageFromBytes(
 		if lastChild != 0 {
 			childVersion := m.pa.GetVersion(lastChild)
 			header := m.pa.GetHeader(pageID)
-			header.extraChild = EncodeChildWithVersion(lastChild, childVersion)
+			header.extraChild = EncodeChildWithVersion(lastChild, uint64(childVersion))
 		} else {
 			header := m.pa.GetHeader(pageID)
 			header.extraChild = 0
